massive realignment
I wasn't satisfied with the existing state of the OAuth oapi docs and
flow, so I tested the whole thing front to back and found/fixed some
issues, notably that the whole Swagger UI was inaccessible, had multiple
spec issues, and did not support logging in through the Authorize flow
the UI usually has for these kinds of APIs.

As of this commit everything is fully OAPI spec compliant, if a bit
scattered, and will import into my Insomnia REST client just fine.

The Swagger UI is now fully accessible without authentication - this was
tough to figure out because of the Spring Security 6 rework to
customization. I eventually got my answer here:
https://spring.io/blog/2022/02/21/spring-security-without-the-websecurityconfigureradapter

Full (but brief) instructions have also been added in the README for
setting up the Swagger UI and a REST client.